Cleanups

+++ src/selector/event.rs
@@ -3,5 +3,10 @@ use crate::selector::Interest;
 #[repr(C, packed(1))]
 pub struct Event {
     events: Interest,
-    token: usize,
+    token: u64,
 }
+
+// struct epoll_event {
+//     uint32_t      events;  /* Epoll events */
+//     epoll_data_t  data;    /* User data variable */
+// };
